Creates mini aggregate span waterfall component for use in webvitals page overview. Move waterfall construction logic to useSpanWaterfallModelFromTransaction from aggregateSpans for reusability
